server/urls.py
from django.conf import settings
from django.contrib import admin
from django.urls import include, path

from server.apps.api import urls as api_urls
from server.apps.dictionaries import urls as dict_urls

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(api_urls)),
    path('refbooks/', include(dict_urls)),
]

if settings.DEBUG:  # pragma: no cover
    import debug_toolbar  # noqa: WPS433
    from django.conf.urls.static import static  # noqa: WPS433
    urlpatterns = [
        # URLs specific only to django-debug-toolbar:
        path('__debug__/', include(debug_toolbar.urls)),
        *urlpatterns,
        # Serving media files in development only:
        *static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT),  # type: ignore[list-item]
    ]
