pyproject.toml
[tool.poetry]
name = "terminology-service"
version = "0.1.0"
description = ""
authors = ["Sidorenkov Anton <asid.owner@gmail.com>"]
packages = [{ include = "server" }]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "3.14.0"
django-split-settings = "1.2"
django-filter = "22.1"
python-decouple = "3.8"
pytz = "2022.7.1"
gunicorn = "20.1.0"
dj-database-url = "1.2.0"
drf-spectacular = "^0.26.0"
structlog = "^22.3.0"
django-stubs-ext = "^0.7.0"
whitenoise = "^6.4.0"
psycopg2 = "^2.9.5"

[tool.poetry.group.dev.dependencies]
ipdb = "0.13.11"
ipython = "8.9.0"
mock = "5.0.1"
django-nose = "1.4.7"
coverage = "7.1.0"
factory-boy = "3.2.1"
django-debug-toolbar = "3.8.1"

wemake-python-styleguide = "^0.17"
flake8-pytest-style = "^1.6"
flake8-django = "^1.1"
flake8-logging-format = "^0.9"
doc8 = "^1.0"

pytest = "^7.1"
pytest-django = "^4.5"
pytest-cov = "^4.0"
django-coverage-plugin = "^2.0"
pytest-randomly = "^3.12"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.54"

safety = "2.1"
mkdocs = "1.4.2"

mypy = "1.0.1"
django-stubs = "1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

